+++ examples/SX128x_examples/Diagnostic_and_Test/11_LoRa_Packet_Logger_Receiver/11_LoRa_Packet_Logger_Receiver.ino
@@ -33,7 +33,7 @@ uint32_t RXpacketCount;
 uint32_t errors;
 
 uint8_t RXPacketL;                                //stores length of packet received
-int8_t  PacketRSSI;                               //stores RSSI of received packet
+int16_t  PacketRSSI;                              //stores RSSI of received packet
 int8_t  PacketSNR;                                //stores signal to noise ratio of received packet
 
 
@@ -220,4 +220,3 @@ void setup()
   Serial.print(F(" Receiver ready"));
   Serial.println();
 }
-
