+++ examples/SX128x_examples/Sensor/18_Sensor_Receiver/18_Sensor_Receiver.ino
@@ -1,21 +1,21 @@
 /*******************************************************************************************************
   Programs for Arduino - Copyright of the author Stuart Robinson - 29/02/20
-  
+
   This program is supplied as is, it is up to the user of the program to decide if the program is
-  suitable for the intended purpose and free from errors. 
+  suitable for the intended purpose and free from errors.
 *******************************************************************************************************/
 
 /*******************************************************************************************************
   Program Operation - The program receives a LoRa packet without using a processor buffer, the LoRa devices
-  internal buffer is read direct for the received sensor data. 
-  
+  internal buffer is read direct for the received sensor data.
+
   The sensor used in the matching '17_Sensor_Transmiter' program is a BME280 and the pressure, humidity,
   and temperature are being and received. There is also a 16bit value of battery mV and and a 8 bit status
   value at the end of the packet.
 
   When the program starts, the LoRa device is setup to set the DIO1 pin high when a packet is received. When
   a packet is received, its printed and assuming the packet is validated, the sensor results are printed to
-  the serial monitor and screen. 
+  the serial monitor and screen.
 
   For the sensor data to be accepted as valid the folowing need to match;
 
@@ -28,8 +28,8 @@
 
   The pin definitions, LoRa frequency and LoRa modem settings are in the Settings.h file.
 
-  With a standard Arduino Pro Mini and SSD1306 display the current consumption was 20.25mA with the 
-  display and 16.6mA without the display.  
+  With a standard Arduino Pro Mini and SSD1306 display the current consumption was 20.25mA with the
+  display and 16.6mA without the display.
 
   Serial monitor baud rate is set at 9600.
 *******************************************************************************************************/
@@ -51,7 +51,7 @@ uint32_t RXpacketErrors;         //count of all packets with errors received
 bool packetisgood;
 
 uint8_t RXPacketL;               //length of received packet
-int8_t  PacketRSSI;              //RSSI of received packet
+int16_t  PacketRSSI;             //RSSI of received packet
 int8_t  PacketSNR;               //signal to noise ratio of received packet
 
 uint8_t RXPacketType;
@@ -121,7 +121,7 @@ void packet_Received_OK()
   printreceptionDetails();                   //print details of reception, RSSI etc
   Serial.println();
 
-  contenterrors = checkPacketValid(len);     //pass length of packet to check routine 
+  contenterrors = checkPacketValid(len);     //pass length of packet to check routine
 
   if (contenterrors == 0)
   {
@@ -129,18 +129,18 @@ void packet_Received_OK()
     ValidPackets++;
     printSensorValues();                     //print the sensor values
     Serial.println();
-    printPacketCounts();                     //print count of valid packets and errors                  
+    printPacketCounts();                     //print count of valid packets and errors
     displayscreen1();
     Serial.println();
   }
   else
   {
-  Serial.println(F("  Packet is not valid"));
-  RXpacketErrors++;
-  disp.clearLine(7);
-  disp.setCursor(0, 7);
-  disp.print(F("Errors "));
-  disp.print(RXpacketErrors);  
+    Serial.println(F("  Packet is not valid"));
+    RXpacketErrors++;
+    disp.clearLine(7);
+    disp.setCursor(0, 7);
+    disp.print(F("Errors "));
+    disp.print(RXpacketErrors);
   }
 }
 
@@ -148,7 +148,7 @@ void packet_Received_OK()
 uint8_t checkPacketValid(uint8_t len)
 {
   //this function checks if the packet is valid and will be displayed
-  
+
   uint8_t errors = 0;
 
   if (RXPacketType != Sensor1)                        //is it a Sensor1 type packet
@@ -156,12 +156,12 @@ uint8_t checkPacketValid(uint8_t len)
     errors++;
   }
 
-  if (RXDestination != This_Node)                     //was the packet sent to this receiver node ? 
+  if (RXDestination != This_Node)                     //was the packet sent to this receiver node ?
   {
     errors++;
   }
 
-  if (!checkCRCvalue(len))                            //is the sent CRC value of sensor data valid ? 
+  if (!checkCRCvalue(len))                            //is the sent CRC value of sensor data valid ?
   {
     errors++;
   }
@@ -179,8 +179,8 @@ bool checkCRCvalue(uint8_t len)
 
   Serial.print(F("len = "));
   Serial.println(len);
-  
-  CRCSensorData = LT.CRCCCITTSX(3, (len-1), 0xFFFF);    //calculate the CRC of packet sensor data
+
+  CRCSensorData = LT.CRCCCITTSX(3, (len - 1), 0xFFFF);  //calculate the CRC of packet sensor data
 
   Serial.print(F("(CRC of Received sensor data "));
   Serial.print(CRCSensorData, HEX);
@@ -231,7 +231,7 @@ void printreceptionDetails()
   Serial.print(F("dBm,SNR,"));
   Serial.print(PacketSNR);
   Serial.print(F("dB,Length,"));
-  Serial.print(LT.readRXPacketL()); 
+  Serial.print(LT.readRXPacketL());
 }
 
 
@@ -240,7 +240,7 @@ void printPacketCounts()
   Serial.print(F("ValidPackets,"));
   Serial.print(ValidPackets);
   Serial.print(F(",Errors,"));
-  Serial.print(RXpacketErrors);  
+  Serial.print(RXpacketErrors);
 }
 
 
@@ -355,4 +355,3 @@ void setup()
   Serial.println(F("Receiver ready"));
   Serial.println();
 }
-
